macos: support build on macOS (incl. s3 support)
requires homebrew. marked as experimental in changelog because of
limited testing and because some usually common options like
"--direct" (O_DIRECT) are not available on MacOS. however,

this also bumps the version number.

+++ source/toolkits/offsetgen/OffsetGenRandomAlignedFullCoverageV2.h
@@ -278,7 +278,7 @@ class OffsetGenRandomAlignedFullCoverageV2 : public OffsetGenerator
             { return blockSize; }
 
         virtual size_t getNextBlockSizeToSubmit() const override
-            { return std::min(numBytesLeft, blockSize); }
+            { return std::min(numBytesLeft, (uint64_t)blockSize); }
 
         virtual uint64_t getNumBytesTotal() const override
             { return numBytesTotal; }
